requirements.txt
gunicorn==20.0.4
async-generator==1.10
backcall==0.2.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-speech==2.0.1
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.15
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.3
notebook==6.0.3
numpy==1.19.4
parso==0.7.0
playsound==1.2.2
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
rsa==4.6
scipy==1.5.4
Send2Trash==1.5.0
terminado==0.9.1
testpath==0.4.4
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Wave==0.0.2
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2